New changes! Woohoogit add -A

+++ db/migrate/20220222020922_create_products.rb
@@ -3,7 +3,7 @@ class CreateProducts < ActiveRecord::Migration[7.0]
     create_table :products do |t|
       t.string  :name
       t.text    :description
-      t.numeric :price
+      t.float :price
       t.integer :stock
 
       t.timestamps
